requirements.txt
Flask==1.1.1 
Flask-JWT-Extended==3.24.0
PyJWT==1.7.1
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.9 
Flask-Migrate==2.5.2  
dynaconf==2.2.1
passlib==1.7.2 
marshmallow==3.3.0
marshmallow-enum==1.5.1
requests==2.22.0 
Flask-Admin==1.5.4
Flask-Bootstrap==3.3.7.1
SQLAlchemy-serializer==1.3.4.2
flask-simplelogin==0.0.7 
mysqlclient==1.4.6
python-telegram-bot    